refactor: Refactor Navigation.kt into per-feature NavGraphBuilder extensions
  Split the 828-line monolithic Navigation.kt into 12 focused extension
  files (one per feature), reducing the orchestrator to ~75 lines. Each
  feature now owns its own composable route wiring
