+++ composeApp/src/commonMain/kotlin/com/thejawnpaul/gptinvestor/core/navigation/TrendingNavGraph.kt
@@ -1,0 +1,11 @@
+package com.thejawnpaul.gptinvestor.core.navigation
+
+import androidx.navigation.NavGraphBuilder
+import androidx.navigation.NavHostController
+import androidx.navigation.compose.composable
+
+fun NavGraphBuilder.trendingNavGraph(navController: NavHostController) {
+    composable(route = Screen.AllTrendingScreen.route) {
+        // TODO: implement all trending screen
+    }
+}
